english-educator-agent/requirements-dev.txt
# Development Dependencies
# Install with: pip install -r requirements-dev.txt

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
faker==22.2.0
hypothesis==6.98.1

# Code Quality
black==24.1.1
ruff==0.1.14
isort==5.13.2
mypy==1.8.0
pylint==3.0.3

# Type Stubs
types-redis==4.6.0.20240106
types-requests==2.31.0.20240125

# Development Tools
ipython==8.20.0
ipdb==0.13.13
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6
mkdocstrings==0.24.0
mkdocstrings-python==1.8.0

# Performance & Profiling
locust==2.20.0
py-spy==0.3.14
memory-profiler==0.61.0

# API Testing
httpx==0.26.0
websockets==12.0

# Database Tools
alembic==1.13.1
sqlalchemy-utils==0.41.1

# Utilities
python-dotenv==1.0.0
watchdog==4.0.0
rich==13.7.0
